requirements.txt
cffi==1.14.1
flake8==3.8.3
importlib-metadata==1.7.0
keyboard==0.13.5
mccabe==0.6.1
miniaudio==1.32
numpy==1.19.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-xlib==0.27
simpleaudio==1.0.4
six==1.15.0
sox==1.4.0
zipp==3.1.0
